% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_timing_data_fct_display.R
\docType{data}
\name{chart_cols}
\alias{chart_cols}
\title{Color palette for plots}
\format{
An object of class \code{character} of length 9.
}
\usage{
chart_cols
}
\description{
Color palette for plots
}
\keyword{datasets}
