% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboard_fct_display.R
\name{valueBox3}
\alias{valueBox3}
\title{Custom value box}
\usage{
valueBox3(
  value,
  title,
  sparkobj = NULL,
  subtitle,
  icon = NULL,
  color = "aqua",
  width = 4,
  href = NULL
)
}
\description{
Custom value box
}
